% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acis-package.R
\docType{package}
\name{acis-package}
\alias{acis}
\alias{acis-package}
\title{acis: R interface to Applied Climate Information System (ACIS) Web Services}
\description{
Provides wrapper functions for Applied Climate Information System (ACIS) Web Services. ACIS Web Services (described at \url{http://www.rcc-acis.org/docs_webservices.html}) are accessed by seven types of calls: station metadata (StnMeta), raw or summarized climate data for a single station for a range of dates (StnData), data for multiple stations (MultiStnData), gridded data sets (GridData & GridData2), hourly data (StnHourly), and metadata for a geographic area (General).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/acis}
  \item Report bugs at \url{https://github.com/brownag/acis/issues}
}

}
\keyword{internal}
