% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{acis_query}
\alias{acis_query}
\alias{acis_call_names}
\title{ACIS Query Interface}
\usage{
acis_query(
  call,
  params,
  base_url = "http://data.rcc-acis.org",
  query_string = FALSE
)

acis_call_names()
}
\arguments{
\item{call}{One of \code{"StnMeta"}, \code{"StnData"}, \code{"MultiStnData"}, \code{"GridData"}, \code{"GridData2"}, \code{"StnHourly"}, \code{"General"}}

\item{params}{A named \emph{list} of parameter values. For complex/nested inputs \code{params} can be a list with 1 element named \code{"params"} containing JSON input as character vector.}

\item{base_url}{Default: \code{"http://data.rcc-acis.org"}}

\item{query_string}{Return query URL rather than result? Default: \code{FALSE}}
}
\value{
\code{acis_query()} returns result of downloading URL, \emph{try-error} on error. Unless otherwise specified in \code{params} the result is JSON (converted to \emph{list}). Output types other than JSON include \code{output="csv"} (returned as \code{data.frame}) and \code{output="image"} (PNG returned as \emph{SpatRaster}).

\code{acis_call_names()} returns a character vector of available calls
}
\description{
\code{acis_query()} builds URLs and makes requests to obtain data from the ACIS Web Services.
}
